LPS-162507 Move initialDeploymentUpgradeStep to ClientExtensionUpgradeStepRegistrator in order to be able to execute upgrade steps either on first install or when new upgrade steps are added.
