+++ modules/apps/client-extension/client-extension-service/src/main/java/com/liferay/client/extension/internal/upgrade/registry/ClientExtensionUpgradeStepRegistrator.java
@@ -16,24 +16,52 @@ package com.liferay.client.extension.internal.upgrade.registry;
 
 import com.liferay.client.extension.internal.upgrade.v3_0_0.ClassNamesUpgradeProcess;
 import com.liferay.client.extension.internal.upgrade.v3_1_0.util.ClientExtensionEntryRelTable;
+import com.liferay.portal.kernel.model.Release;
+import com.liferay.portal.kernel.service.ReleaseLocalService;
 import com.liferay.portal.kernel.upgrade.BaseExternalReferenceCodeUpgradeProcess;
 import com.liferay.portal.kernel.upgrade.CTModelUpgradeProcess;
+import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.upgrade.UpgradeProcessFactory;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Iván Zaera
  */
-@Component(
-	enabled = false, immediate = true, service = UpgradeStepRegistrator.class
-)
+@Component(immediate = true, service = UpgradeStepRegistrator.class)
 public class ClientExtensionUpgradeStepRegistrator
 	implements UpgradeStepRegistrator {
 
 	@Override
 	public void register(Registry registry) {
+		registry.registerInitialDeploymentUpgradeSteps(
+			new UpgradeProcess() {
+
+				@Override
+				protected void doUpgrade() throws Exception {
+					Release remoteAppRelease =
+						_releaseLocalService.fetchRelease(
+							"com.liferay.remote.app.service");
+
+					if (remoteAppRelease != null) {
+						Release clientExtensionRelease =
+							_releaseLocalService.fetchRelease(
+								"com.liferay.client.extension.service");
+
+						clientExtensionRelease.setSchemaVersion(
+							remoteAppRelease.getSchemaVersion());
+
+						_releaseLocalService.updateRelease(
+							clientExtensionRelease);
+
+						_releaseLocalService.deleteRelease(remoteAppRelease);
+					}
+				}
+
+			});
+
 		registry.register(
 			"1.0.0", "1.0.1",
 			UpgradeProcessFactory.alterColumnType(
@@ -108,5 +136,8 @@ public class ClientExtensionUpgradeStepRegistrator
 				"ClientExtensionEntryRel", "groupId LONG",
 				"lastPublishDate DATE null"));
 	}
+
+	@Reference
+	private ReleaseLocalService _releaseLocalService;
 
 }
